refactor code and fixed sshuttle version check error
