+++ pkg/kt/command/types.go
@@ -1,0 +1,10 @@
+package command
+
+import (
+	"github.com/alibaba/kt-connect/pkg/kt/options"
+)
+
+// Action cmd action
+type Action struct {
+	Options *options.DaemonOptions
+}
